Refactor AutoDateTimeField out of management.utils
