+++ rbac/management/utils.py
@@ -103,11 +103,3 @@ def queryset_by_id(objects, clazz):
     """Return a queryset of from the class ordered by id."""
     wanted_ids = [obj.id for obj in objects]
     return clazz.objects.filter(id__in=wanted_ids).order_by('id')
-
-
-class AutoDateTimeField(models.DateTimeField):
-    """Class that defines is pre_save value."""
-
-    def pre_save(self, model_instance, add):
-        """Save its time as now."""
-        return timezone.now()
